refactor code to use functions from parser
