+++ pkg/ingress2gateway/translate/translate_action_helper_test.go
@@ -12,112 +12,6 @@ import (
 	gwv1 "sigs.k8s.io/gateway-api/apis/v1"
 )
 
-func TestParseActionAnnotation(t *testing.T) {
-	tests := []struct {
-		name    string
-		annos   map[string]string
-		svcName string
-		wantErr bool
-		check   func(t *testing.T, a *ingress.Action)
-	}{
-		{
-			name: "fixed-response action",
-			annos: map[string]string{
-				"alb.ingress.kubernetes.io/actions.response-503": `{"type":"fixed-response","fixedResponseConfig":{"contentType":"text/plain","statusCode":"503","messageBody":"503 error text"}}`,
-			},
-			svcName: "response-503",
-			check: func(t *testing.T, a *ingress.Action) {
-				assert.Equal(t, ingress.ActionTypeFixedResponse, a.Type)
-				require.NotNil(t, a.FixedResponseConfig)
-				assert.Equal(t, "503", a.FixedResponseConfig.StatusCode)
-				assert.Equal(t, "503 error text", *a.FixedResponseConfig.MessageBody)
-			},
-		},
-		{
-			name: "redirect action",
-			annos: map[string]string{
-				"alb.ingress.kubernetes.io/actions.redirect-to-eks": `{"type":"redirect","redirectConfig":{"host":"aws.amazon.com","path":"/eks/","port":"443","protocol":"HTTPS","query":"k=v","statusCode":"HTTP_302"}}`,
-			},
-			svcName: "redirect-to-eks",
-			check: func(t *testing.T, a *ingress.Action) {
-				assert.Equal(t, ingress.ActionTypeRedirect, a.Type)
-				require.NotNil(t, a.RedirectConfig)
-				assert.Equal(t, "aws.amazon.com", *a.RedirectConfig.Host)
-				assert.Equal(t, "HTTP_302", a.RedirectConfig.StatusCode)
-			},
-		},
-		{
-			name: "forward single TG by ARN (simplified schema)",
-			annos: map[string]string{
-				"alb.ingress.kubernetes.io/actions.forward-single-tg": `{"type":"forward","targetGroupARN":"arn:aws:elasticloadbalancing:us-west-2:123456789:targetgroup/my-tg/abc123"}`,
-			},
-			svcName: "forward-single-tg",
-			check: func(t *testing.T, a *ingress.Action) {
-				assert.Equal(t, ingress.ActionTypeForward, a.Type)
-				require.NotNil(t, a.ForwardConfig)
-				require.Len(t, a.ForwardConfig.TargetGroups, 1)
-				require.NotNil(t, a.ForwardConfig.TargetGroups[0].TargetGroupARN)
-				assert.Equal(t, "arn:aws:elasticloadbalancing:us-west-2:123456789:targetgroup/my-tg/abc123", *a.ForwardConfig.TargetGroups[0].TargetGroupARN)
-			},
-		},
-		{
-			name: "forward single TG by name (simplified schema)",
-			annos: map[string]string{
-				"alb.ingress.kubernetes.io/actions.forward-by-name": `{"type":"forward","targetGroupName":"my-target-group"}`,
-			},
-			svcName: "forward-by-name",
-			check: func(t *testing.T, a *ingress.Action) {
-				assert.Equal(t, ingress.ActionTypeForward, a.Type)
-				require.NotNil(t, a.ForwardConfig)
-				require.Len(t, a.ForwardConfig.TargetGroups, 1)
-				require.NotNil(t, a.ForwardConfig.TargetGroups[0].TargetGroupName)
-				assert.Equal(t, "my-target-group", *a.ForwardConfig.TargetGroups[0].TargetGroupName)
-			},
-		},
-		{
-			name: "forward multiple TGs with weights and stickiness (advanced schema)",
-			annos: map[string]string{
-				"alb.ingress.kubernetes.io/actions.forward-multiple-tg": `{"type":"forward","forwardConfig":{"targetGroups":[{"serviceName":"service-1","servicePort":"http","weight":20},{"serviceName":"service-2","servicePort":80,"weight":20},{"targetGroupARN":"arn-of-your-non-k8s-target-group","weight":60}],"targetGroupStickinessConfig":{"enabled":true,"durationSeconds":200}}}`,
-			},
-			svcName: "forward-multiple-tg",
-			check: func(t *testing.T, a *ingress.Action) {
-				assert.Equal(t, ingress.ActionTypeForward, a.Type)
-				require.NotNil(t, a.ForwardConfig)
-				require.Len(t, a.ForwardConfig.TargetGroups, 3)
-				require.NotNil(t, a.ForwardConfig.TargetGroupStickinessConfig)
-				assert.Equal(t, true, *a.ForwardConfig.TargetGroupStickinessConfig.Enabled)
-				assert.Equal(t, int32(200), *a.ForwardConfig.TargetGroupStickinessConfig.DurationSeconds)
-			},
-		},
-		{
-			name:    "missing annotation",
-			annos:   map[string]string{},
-			svcName: "nonexistent",
-			wantErr: true,
-		},
-		{
-			name: "invalid JSON",
-			annos: map[string]string{
-				"alb.ingress.kubernetes.io/actions.bad": `{invalid`,
-			},
-			svcName: "bad",
-			wantErr: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			a, err := parseActionAnnotation(tt.annos, tt.svcName)
-			if tt.wantErr {
-				require.Error(t, err)
-				return
-			}
-			require.NoError(t, err)
-			tt.check(t, a)
-		})
-	}
-}
-
 func TestTranslateAction_FixedResponse(t *testing.T) {
 	a := &ingress.Action{
 		Type: ingress.ActionTypeFixedResponse,
